Snippet of JS to tap into Nexus Telnet
Some JavaScript to hijack the websocket prototype send function and
capture any websocket that uses it.

Also opens up a websocket to the Erlang server to forward all data
there.

Has some convenience code for changing the focus to the input line as
well. This has nothing to do with this GUI layer on top of Nexus.