Subscriber Web Page for Nexus Data
This web page simply opens a websocket connection to the local Erlang
server and registers for a type of MUD data. Currently the only type of
data is "raw".

Once the Erlang server can build a model of the Nexus MUD data I can
start to add different data types, e.g.:
- map
- HP
- Gold
- Inventory
- Custom room notes
- Time
- Who
- etc.