Parsing generally works now
I had to fix all kinds of things to get the parsing to work:
- Telnet feature will/wont (log and ignore)
- Telnet feature set to value (log and ignore)
- GMCP message ends with no ?IAC ?SE terminator
- Various colors in the 38;5;0xxm telnet control code
- Color reset on logout (haven't tested this)
- etc.

Add new subscription data types.

Send subscriptions to everyone subscribed and to anyone subscribed to
raw. If a process subscribes to both they'll get dupes.

Default the subscription lists so I don't have to check for empty lists.

Convert incoming subscription types to atoms manually (don't want to
create atoms on the fly) so that I don't have to deal with data types in
mudspy_parse.erl